Current pico 8 cart i have
This is what I currently have for Carne Castle.  I reduced the number of tokens used for player movement, however the bug where you move all the way to the right most position of the matrix is still an issue. I put in some basic wall and door sprites, and put the walls into GB.