Firestarter and Fire
I added the Firestarter sprites, animation, and AI.  It seems to be working as intended.  I also added in fire and updated everything but the player to interact with the fire.

The bug with the sprite in the upper-left corner now sometimes shows fire - this bug has me confused.  Let me know if y'all have any ideas.